requirements.txt
sentence-transformers
faiss-cpu
gpt4all
PyPDF2
streamlit
numpy
scikit-learn
